Windows XRT: Address several windows compiler warnings / issues (#1485)
+ Updated code to promote key integers to uint64_t
    + Renamed XclBin.h/cxx to XclBinClass.h/cxx to account for case insensitive.
    + Structure memory initialization updated to support the windows compiler

+++ src/runtime_src/tools/xclbin/DTCStringsBlock.cxx
@@ -91,7 +91,7 @@ DTCStringsBlock::addString(const std::string _dtcString)
 
   // Did we find it?
   if (index != std::string::npos) {
-    return index;
+    return (unsigned int) index;
   }
 
   // Not found, lets add it
@@ -99,7 +99,7 @@ DTCStringsBlock::addString(const std::string _dtcString)
 
   *m_pDTCStringBlock << _dtcString << '\0';
  
-  return index;
+  return (unsigned int) index;
 }
